Full update of all input and output data
This is a pretty major release as it moves from the Alexa Top 1M to Tranco. As such, the input data is quite different and that's changed the output data which subsequently appears on the website. The launch blog post for this release explains more: https://www.troyhunt.com/still-why-no-https/